[electrum] fix a few derivation path quirks in testnet mode
Summary:
When running in testnet mode, the long information message about the choice of derivation paths when creating a new wallet is irrelevant. So don't print it.

Also hardcode the paths in the derivation path scanner. The `bip44_derivation_*` functions check for testnet mode to decide which path to return, and in testnet mode they all return the same path. So we ended up scanning the `m/44'/1'/0'`  path 3 times and we don't scan for the 899', the 145' and the 0' paths. Also in mainnet mode we never scanned the tesnet path. Always scan all of these paths.

Test Plan:
`./electrum-abc --testnet`
Restore a wallet from seed. Try the derivation path scanner.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18676