[avalanche] Fix abc_rpc_getavalancheinfo.py flakiness
Summary:
The end of the test invalidates the local proof while voting unknown for the other items. But there are more than 16 items to be polled (16 proofs for the quorum, 1 block and the local proof) so if the local proof is not part of the initial 16 items being polled, proofs from the quorum might be stalled and the quorum lost before the local proof had a chance of being polled.

Instead we just finalize everything but the local proof and adjust the expected finalized proof count accordingly.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getavalancheinfo

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18694