[avalanche] Gate the activation of avalanche preconsensus on november 15
Summary:
This will only activate if enabled and if the upgrade is activated. The same applies to mining with avalanche preconsensus which has its own flag.

In order to make it easier to test, the getinfo RPC is extended to report the activation status.

Depends on D18689 and D18700.

Test Plan:
  ninja check check-functional-extended
  ninja check-upgrade-activated

  # Check BlockAssemblerAddPackageTxnsPreconsensus is faster than
  # BlockAssemblerAddPackageTxns
  ninja bench-bitcoin

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18701