[Cashtab] Patch bug in tx history parsing
Summary:
`wallets[0]` is no longer the active wallet, as we no longer modify the order of the stored `wallets` every time the user changes a wallet. Access this variable appropriately.

Note: without this fix, we were parsing tx notifications for the (probably) wrong wallet, as a user with multiple wallets is unlikely to have the active wallet be wallets[0].

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18755