[e.cash] Use Strapi CMS for build page products
Summary:
Was requested to migrate the data for the "Built with XEC" product list on the /build page to strapi.
Hopefully this list is ever changing, so will be nice to be able to make edits to it easily through strapi as needed.

So adding some fetch logic similar to the blog to get the new strapi endpoint for products.

Test Plan: preview and check. Should look the same as current but data now coming from strapi

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18790