[Cashtab] Initial ios support
Summary:
This adds support for building and running Cashtab as an IOS app.
There are still some issues to fix, especially the safe area insets (on devices with a notch, the app collides with the notification bar) but this is a good starting point.

Test Plan:
  ./start-ios.sh
Check on a real device that everything works as expected, including the qr scanner.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman, johnkuney

Differential Revision: https://reviews.bitcoinabc.org/D18802