[Cashtab] IOS: play nicely with the notification bar
Summary:
On devices with a notch the app collides with the notification area. This is fixed by offsetting the content using the safe area values provided by iOS WebKit.
This leaves the notification bar with the default device theme which is ugly, so we also override the theme for this app to match Cashtab design.

Before:
{F18712310}

After:
{F18712382}

Test Plan:
  ./start-ios.sh
Check it doesn't break anything on Android:
  ./start-android.sh

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D18804