[cashtab, ecash-agora] Support for 64bit ints in agora offers
Summary:
`ecash-agora` already has support for 64bit ints, though this needs to be exposed in a param for apps to access it.

Expose this and version bump in ecash-agora.

Implement in Cashtab, gated by a param.

This diff also makes some corrections to the use of testnet env vars that were necessary to get a functional testnet instance running locally.

Test Plan: `npm test`, create and buy agora offers on testnet at cashtab.io, which has this diff deployed on testnet

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18829