[ecash-wallet] Patch ts issue to stay consistent with defined interface
Summary:
If an app uses `ecash-wallet` as a dep and has strict type checking, it will throw an error on this.

Properly speaking we only want to have a `missingTokens` key if there are missing tokens; we do not really want `missingTokens: undefined`

Patch.

Test Plan: `npm test` and CI tests

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18839