[ecash-agora] Specify ecash-wallet as a dep and not a dev dep
Summary:
Missed this after adding methods to actual lib methods that depend on ecash-wallet. CI was still passing as it was getting the deb. But third party apps that install ecash-agora are not getting the ecash-wallet dep and they should be.

Move it to deps to patch.

Test Plan: `npm test`, CI, publish and test install in third party apps

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18852