refactor: Replace GetTimeMicros by SystemClock
Summary:
```
It is unclear from the name that GetTimeMicros returns the system time. Also, it is not using the type-safe std::chrono types.

Fix both issues by replacing it with SystemClock in the only place it is used.

This refactor should not change behavior.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/27233 | core#27233]].

Depends on D18873.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18874