[metachronik] Another railway CI fix attempt
Summary:
Changed the failure msg with the last diff. Now getting some kind of ssl cert errors. I am not sure if this has to do with the image or not. This fix was suggested by AI.

If this does not work, I will get it to just run on one of our servers instead of railway, or remove the CI.

Test Plan:
`docker build -f metachronik.Dockerfile -t metachronik-test .` runs, build works

land and check CI, I can't reproduce this locally as it seems to be a server / railway auth issue.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18897