[metachronik] Remove CI and document prod deployments
Summary:
Not worth fighting the mysterious CI issues for the railway deployments. This is simple enough to deploy locally or through the railway interface. In practice, many deployments are often necessary to test different indexing.

Add documentation. The app is still production ready for alternative deployment solutions if Railway changes or it otherwise becomes unmaintained.

Test Plan: proofread. metachronik CI is deleted

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18901