[lint] enable FLY002 checks for the Electrum codebase
Summary:
Checks for str.join calls that can be replaced with f-strings for better readability.

This check is easy to enable, so let's do it immediately to reduce the style divergence between electrum and the rest of the codebase.
The other two (convert % format and `.format` into f-strings) are a bit more tedious and will be done in multiple chunks

Test Plan:
```
arc lint --everything
ninja check-electrum check-electrum-functional

```

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18917