[ecash-lib] Export new constants and interfaces to be used for chained txs in ecash-wallet
Summary:
Ref D18867

Isolate the ecash-lib changes needed for ecash-wallet chained txs in this diff. In this way, we can ensure they are published and available for ecash-wallet to pull in as a dep.

The changes to action.ts introduce optional params not yet supported in `ecash-wallet`. The published version of `ecash-wallet` will not be on this version of `ecash-lib` until D18867 lands, so no behavior change is expected.

Test Plan: `npm test`, CI

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18918