[ecash-wallet] Remove consumed postage utxos in addFuelAndSign
Summary:
Match the same API from `build()`

Note `buildPostage` does not update the utxo set, because it is not necessarily possible to know the txid and the change utxos that go back to the postage wallet.

However, we should do this for the fuel wallet in addFuelAndSign, because in this case we know that there are only consumed utxos and no change utxos. There are significant server speed and resource improvements available for postage payment servers with this update.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18924