[ecash-agora] Patch version bump for latest ecash-wallet dep
Summary:
It's much better for users with all of these libs in the same app if all of the peer deps are also the latest version

Argument that ... should just have one lib. But I think the value of modularity still outweighs the inconvenience of these maintenace updates.

Test Plan: CI tests

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18925