[cashtab-faucet] Tweak some recaptcha settings to see if it slows drain attacks
Summary:
Consistent hits, esp on the 42-XEC rewards.

It's actually a pretty great test of how little it must cost for someone to get a 0.7 recaptcha score and rotate IP addresses.

Ultimately we should require a bit more and offer a bit more in the faucet. But have not thought this through yet. For now, see what happens if we require a higher recaptcha score.

Because I know that this score is often missed even in manual attemps, we also bump the number of acceptable attempts.

Test Plan: land and monitor. Will prob get some complaints in the tg about it not working. But main thing to check is if the attacker keeps it up.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18926