[electrum] fix unit for proof editor
Summary:
The "select coins in wallet" widget showed "sats" when the amounts where actually formatted as "XEC". Let's just respect the unit selected by the user (default XEC).
Print whitespaces in case the user has not selected full precision in the settings (e.g unit XEC and num_zeros 0 or 1) so that the integer parts align vertically in the amount column, the same way it is already done in the Coins tab.

Depends on D18988

Test Plan: open the proof editor via the menu and via the coins tab (context menu when clicking on a utxo). Try to load additional coins from the wallet.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18989