[avalanche] Reduce the logs verbosity
Summary: Again. Deployment on mainnet did show that printing unconditionally the finalization message for each tx is quite spammy. Similarly most users are not interested in the stake contenders vote. We move these logs to the avalanche log category so they become opt-in.

Test Plan:
  ninja all check-all
Run a node and check the log are less verbose.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18998