[marlin] Add Marlin Wallet to the monorepo - web
Summary:
This is the initial diff for moving Marlin Wallet from its github to the monorepo.

This diff adds the web version of the wallet, which can be run as is but will lack all the mobile/hardware features. The other files will be added in follow ups so it's easier to review.

The files are similar to the github source (https://github.com/Fabcien/MarlinWallet) but not identical:
 - they have a proper copyright header.
 - they have been linted according to the monorepo rules.

There is no test nor CI for now (as in the source repo) but can be added later.

Test Plan: Read the README and try the instructions.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D19047