[Cashtab] Show spinner while importing wallet
Summary:
Importing a wallet with many tokens can take some time from the Onboarding screen, as we do not navigate to the wallet screen until it is loaded.

We should disable the button and show a spinner to indicate that something is happening, otherwise it seems like the Import feature is broken.

Done in this diff.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19117