[bitcoinsuite] Fix tests; also run tests when Chronik changes
Summary:
Currently, the tests for bitcoinsuite-chronik-client don't build, because chronik.proto changed and the Python tests in `setup_scripts` were updated a while back, but this remained undetected.

This diff fixes the test, and also re-runs the bitcoinsuite-chronik-client tests whenever Chronik or the Python tests change.

The Python tests have some additional behavior now, but implementing this behavior is outside of the scope of this diff, which just aims to fix the build.

Test Plan: `cargo test -p bitcoinsuite-chronik-client`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D19130