[ecash-wallet] Scoping out HD support
Summary:
In order for apps that currently utilize HD wallets to extend token support, need a token-aware HD wallet. ecash-wallet is best positioned for this, and HD support is on the roadmap.

Add a README file for scoping out and planning this support.

The goal is to have feature parity with existing ABC HD wallets, such as Electrum ABC and the in-node wallet.

Test Plan: proofread

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19153