[explorer] Automate caching hash updates
Summary:
This translates the bash script into a rust template replacement.
The downside is that new files need to be added manually, the benefit is that it is now automated and can't be missed anymore.

Test Plan:
  cargo run
Inspect the code and check the hashes are set properly in the returned html.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D19181