[marlin] Make Marlin build to work with pnpm
Summary:
This requires some more dev packages and a couple metro/gradle config changes to accomodate the monorepo strategy.
The most annoying change is figuring out where the hermes binary is located as the path is version dependant.
To avoid future issue a function is defined that is version agnostic.

Test Plan:
  # Remove all existing node_modules so we don't have any cached install
  git clean -xffd

  cd apps/marlin
  pnpm install
  # Below commands assume you did set ANDROID_HOME and have metro running via npx react-native start --reset-cache in another terminal
  pnpm run android
  pnpn run build-android-release-aab

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D19192