[marlin] Bugfix: Don't account for tokens when computing the spendable amount
Summary:
We don't want to burn tokens (and we don't) but the spendable amount is not filtering for them. This causes the slider to show an insufficient balance error when sending too much XEC in the send screen. Note that this is not preventing from sending but makes the slider annoying to use as you need to find the spot that works.

Depends on D19192 for testability.

Test Plan: import a wallet with some tokens and try to send all the money via the slider. It should never show "insufficient balance".

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D19193